refactor: domain 패키지 usecase 위치 변경
