+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/domain/usecase/DeleteDiaryUseCase.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.domain.attendance
+package com.yagubogu.domain.usecase
 
 import com.yagubogu.data.repository.checkin.CheckInRepository
 import kotlinx.coroutines.async
