+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/domain/usecase/LoadDiaryUseCase.kt
@@ -1,6 +1,7 @@
-package com.yagubogu.domain.attendance
+package com.yagubogu.domain.usecase
 
 import com.yagubogu.data.repository.checkin.CheckInRepository
+import com.yagubogu.domain.model.AttendanceDiary
 import kotlinx.coroutines.async
 import kotlinx.coroutines.coroutineScope
 
@@ -14,7 +15,8 @@ class LoadDiaryUseCase(
                 val imagesDeferred = async { checkInRepository.getImages(checkInId).getOrThrow() }
                 AttendanceDiary(
                     memo = memoDeferred.await().orEmpty(),
-                    images = imagesDeferred.await().map { AttendanceDiary.Image(id = it.imageId, url = it.imageUrl) },
+                    images = imagesDeferred.await()
+                        .map { AttendanceDiary.Image(id = it.imageId, url = it.imageUrl) },
                 )
             }
         }
